dlタグのregisterBlockTypeを修正
