+++ src/blocks/definition-list/save.js
@@ -1,0 +1,8 @@
+import { useBlockProps, useInnerBlocksProps } from '@wordpress/block-editor';
+
+export default function Save() {
+	const saveBlockProps = useBlockProps.save();
+	const saveInnerBlocksProps = useInnerBlocksProps.save( saveBlockProps );
+
+	return <dl { ...saveInnerBlocksProps } />;
+}
